src/requirements.txt
yfinance
pandas
python-dotenv
sqlalchemy
pandas
psycopg2-binary